Move sign function into signature module
