+++ src/util/key.rs
@@ -27,11 +27,8 @@ use std::str::FromStr;
 
 use consensus::{encode, Decodable, Encodable};
 use network::constants::Network;
-use secp256k1::{self, Secp256k1, SecretKey};
+use secp256k1::{self, Secp256k1};
 use util::base58;
-use util::signature::Signature;
-use util::prime::jacobi;
-use util::rfc7969::nonce_rfc6979;
 
 /// A key-related error.
 #[derive(Debug)]
@@ -40,8 +37,6 @@ pub enum Error {
     Base58(base58::Error),
     /// secp256k1-related error
     Secp256k1(secp256k1::Error),
-    /// Generation for sign nonce related error
-    NonceGeneration
 }
 
 
@@ -50,7 +45,6 @@ impl fmt::Display for Error {
         match *self {
             Error::Base58(ref e) => write!(f, "base58 error: {}", e),
             Error::Secp256k1(ref e) => write!(f, "secp256k1 error: {}", e),
-            Error::NonceGeneration => write!(f, "nonce generation error"),
         }
     }
 }
@@ -60,7 +54,6 @@ impl error::Error for Error {
         match *self {
             Error::Base58(ref e) => Some(e),
             Error::Secp256k1(ref e) => Some(e),
-            Error::NonceGeneration => None,
         }
     }
 
@@ -228,73 +221,6 @@ impl PrivateKey {
             key: secp256k1::SecretKey::from_slice(&data[1..33])?,
         })
     }
-
-    /// Generate schnorr signature
-    pub fn sign_schnorr(&self, message: &[u8; 32]) -> Result<Signature, Error> {
-        let ctx = secp256k1::Secp256k1::signing_only();
-
-        let pk = secp256k1::PublicKey::from_secret_key(&ctx, &self.key);
-
-        // Generate k
-        let mut k = self.generate_k(message)?;
-
-        // TODO: Check private key and k is not zero
-        // this is no need because all secret key instance checked.
-
-        // Compute R = k * G
-        let r = secp256k1::PublicKey::from_secret_key(&ctx, &k);
-
-        // Negate k if value of jacobi(R.y) is not 1
-        if jacobi(&r.serialize_uncompressed()[33..]) != 1 {
-            k.negate_assign();
-        }
-
-        // Compute e = sha256(R.x, pk, message)
-        let e = Signature::compute_e(&r.serialize()[1..33], &pk, message)?;
-
-        // Compute s = k + ep
-        let sigma = {
-            let mut result = e.clone();
-            result.mul_assign(&self.key[..])?;
-            result.add_assign(&k[..])?;
-            result
-        };
-
-        let mut r_x = [0u8; 32];
-        r_x.clone_from_slice(&r.serialize()[1..33]);
-
-        Ok(Signature { r_x, sigma: to_bytes(&sigma) })
-    }
-
-    fn generate_k(&self, message: &[u8; 32]) -> Result<SecretKey, Error> {
-        // "SCHNORR + SHA256"
-        const ALGO16: [u8; 16] = [
-            83, 67, 72, 78, 79, 82, 82, 32, 43, 32, 83, 72, 65, 50, 53, 54
-        ];
-
-        let mut count: u32 = 0;
-
-        loop {
-            let nonce = nonce_rfc6979(
-                message,
-                &self.key,
-                &ALGO16,
-                None,
-                count
-            );
-            count += 1;
-
-            if let Ok(k) = SecretKey::from_slice(&nonce[..]) {
-                return Ok(k);
-            }
-        }
-    }
-}
-
-fn to_bytes(sk: &secp256k1::SecretKey) -> [u8; 32] {
-    let mut r = [0u8; 32];
-    r.clone_from_slice(&sk[..]);
-    r
 }
 
 impl fmt::Display for PrivateKey {
@@ -458,7 +384,6 @@ mod tests {
     use network::constants::Network::Testnet;
     use network::constants::Network::Bitcoin;
     use util::address::Address;
-    use hashes::Hash;
 
     #[test]
     fn test_key_derivation() {
@@ -565,23 +490,5 @@ mod tests {
 
         let decoded = PublicKey::consensus_decode(&s[..]).unwrap();
         assert_eq!(decoded, pk);
-    }
-
-    #[test]
-    fn test_sign_schnorr() {
-        for n in 0..16 {
-            let msg = {
-                let m = format!("Very secret message {}: 11", n);
-                let hash = hashes::sha256::Hash::hash(m.as_bytes());
-                hash.into_inner()
-            };
-
-            let key = PrivateKey::from_wif("5HxWvvfubhXpYYpS3tJkw6fq9jE9j18THftkZjHHfmFiWtmAbrj").unwrap();
-
-            let sign = key.sign_schnorr(&msg).unwrap();
-
-            let ctx = secp256k1::Secp256k1::signing_only();
-            assert!(sign.verify(&msg[..], &key.public_key(&ctx)).is_ok());
-        }
     }
 }
